Add files via upload
Beacon frame is one of the management frames in IEEE 802.11 based WLANs. It contains all the information about the network. Beacon frames are transmitted periodically, they serve to announce the presence of a wireless LAN and to synchronise the members of the service set. Beacons are used in networking to advertise presence of Access Points to clients who wants to connect to the AP.